refactor: use new FormInDialog component for AdminScoringForm
